requirements/base.txt
django==3.0.9
django-countries==6.1.3
django-environ==0.4.5
django-extensions==2.1.6
django-redis==4.12.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
redis==3.5.3
ujson==3.1.0
Werkzeug==0.15.2